+++ src/chain/src/chain.rs
@@ -43,7 +43,9 @@ where
     fn canonical_top(&self) -> Arc<B>;
 
     /// Returns an atomic reference to the genesis block in the chain.
-    fn genesis(&self) -> Arc<B>;
+    fn genesis() -> Arc<B> {
+        Arc::new(B::genesis())
+    }
 
     /// Attempts to append a new block to the chain.
     fn append_block(&mut self, block: Arc<B>) -> Result<(), ChainErr>;
