+++ src/chain/src/hard_chain/chain.rs
@@ -18,14 +18,16 @@
 
 use crate::block::Block;
 use crate::chain::{Chain, ChainErr};
-use crate::easy_block::EasyBlock;
-use crate::iterators::easy::EasyBlockIterator;
+use crate::easy_chain::chain::EasyChain;
+use crate::hard_chain::block::HardBlock;
+use crate::iterators::hard::HardBlockIterator;
 use bin_tools::*;
 use crypto::Hash;
 use elastic_array::ElasticArray128;
 use hashbrown::HashSet;
 use hashdb::HashDB;
 use lru::LruCache;
+use parking_lot::RwLock;
 use persistence::PersistentDb;
 use std::cell::RefCell;
 use std::sync::Arc;
@@ -34,53 +36,59 @@ use rlp::Rlp;
 /// Size of the block cache.
 const BLOCK_CACHE_SIZE: usize = 20;
 
-/// The easy chain stores blocks that represent buffered
-/// validator pool join requests. If a miner wishes to become
-/// a validator, it will most probably mine on the easy chain 
-/// (which has lower difficulty in order to populate the pool
-/// more effectively).
+/// The hard chain stores blocks that represent state can be
+/// changes in the validator pool. A block from the hard chain
+/// can be thought of as a function which changes the state of
+/// the validator pool.
 ///
-/// The difficulty of the easy chain grows asimptotically with
-/// the number of mined blocks since the last mined block on the
-/// hard-chain so that the buffer is rate-limited.
+/// From the point of view of the validator pool a
+/// block mined on the hard chain represents an
+/// injection of:
 ///
-/// When a block is mined on the hard chain, all of the miners
-/// that have succesfuly mined a block on the easy chain (along
-/// with the miner that succesfuly mined a hard block) since
-/// the last mined block on the hard one are joined to the pool
-/// in one operation.
+/// 1. An additional amount of events that the whole pool can order.
+/// 2. Additional validators.
 ///
-/// Miner rewards on the easy chain are substantially less than the
-/// ones on the hard chain, however, miners from the easy chain receive
-/// transaction fees as additional reward because they participate in the
-/// validator pool.
-pub struct EasyChain {
-    /// Reference to the database storing the `EasyChain`.
+/// The pool cannot start ordering events without a block
+/// being mined in the hard chain which states the new
+/// validators that will be added (miners of the latest
+/// easy chain blocks since that last mined hard block),
+/// how many events the pool can order in the next round,
+/// and what nodes to retire from the pool.
+///
+/// At the same time, the next hard block cannot be applied
+/// to the pool until the pool has either consumed all of
+/// their allocated events or until the pool is deemed to be
+/// corrupt.
+pub struct HardChain {
+    /// Reference to the database storing the `HardChain`.
     db: PersistentDb,
 
-    /// The current height of the canonical chain.
+    /// Reference to associated easy chain.
+    easy_chain: Arc<RwLock<EasyChain>>,
+
+    /// The current height of the chain.
     height: usize,
 
     /// The topmost block in the canonical chain.
-    canonical_top: Arc<EasyBlock>,
+    canonical_top: Arc<HardBlock>,
 
     /// Cache storing the top blocks descended from the
     /// canonical chain and excluding the actual canonical
     /// top block.
-    canonical_tops_cache: HashSet<Arc<EasyBlock>>,
+    canonical_tops_cache: HashSet<Arc<HardBlock>>,
 
     /// Cache storing the top blocks of chains that are
     /// disconnected from the canonical chain.
-    pending_tops_cache: HashSet<Arc<EasyBlock>>,
+    pending_tops_cache: HashSet<Arc<HardBlock>>,
 
     /// Block lookup cache
-    block_cache: RefCell<LruCache<Hash, Arc<EasyBlock>>>,
+    block_cache: RefCell<LruCache<Hash, Arc<HardBlock>>>,
 }
 
-impl EasyChain {
-    pub fn new(mut db_ref: PersistentDb) -> EasyChain {
+impl HardChain {
+    pub fn new(mut db_ref: PersistentDb, easy_chain: Arc<RwLock<EasyChain>>) -> HardChain {
         // TODO: Handle different branches
-        let top_key = crypto::hash_slice(b"canonical_top");
+        let top_key = crypto::hash_slice(b"top");
         let top_db_res = db_ref.get(&top_key);
         let canonical_top = match top_db_res.clone() {
             Some(top) => {
@@ -88,17 +96,14 @@ impl EasyChain {
                 buf.copy_from_slice(&top);
 
                 let block_bytes = db_ref.get(&Hash(buf)).unwrap();
-                Arc::new(EasyBlock::from_bytes(&block_bytes).unwrap())
+                Arc::new(HardBlock::from_bytes(&block_bytes).unwrap())
             }
             None => {
-                // TODO: Compute genesis block
-                Arc::new(EasyBlock::new(None))
+                HardChain::genesis()
             }
         };
 
         let canonical_tops_key = crypto::hash_slice(b"canonical_tops");
-
-        // Insert new canonical tops entry if non-existent.
         let canonical_tops = match db_ref.get(&canonical_tops_key) {
             Some(encoded) => {
                 parse_encoded_blocks(&encoded)
@@ -112,7 +117,7 @@ impl EasyChain {
                     ElasticArray128::<u8>::from_slice(&encoded),
                 );
 
-                let mut b: HashSet<Arc<EasyBlock>> = HashSet::new();
+                let mut b: HashSet<Arc<HardBlock>> = HashSet::new();
                 b.insert(canonical_top.clone());
 
                 b
@@ -131,7 +136,8 @@ impl EasyChain {
             }
         };
 
-        let height_key = crypto::hash_slice(b"canonical_height");
+        // TODO: Handle different branches with different heights
+        let height_key = crypto::hash_slice(b"height");
         let height = match db_ref.get(&height_key) {
             Some(height) => decode_be_u64!(&height).unwrap(),
             None => {
@@ -149,40 +155,37 @@ impl EasyChain {
 
         let height = height as usize;
 
-        EasyChain {
+        HardChain {
             canonical_top,
             canonical_tops_cache: canonical_tops,
             pending_tops_cache: pending_tops,
             height,
+            easy_chain,
             db: db_ref,
             block_cache: RefCell::new(LruCache::new(BLOCK_CACHE_SIZE)),
         }
     }
 }
 
-fn parse_encoded_blocks(bytes: &[u8]) -> HashSet<Arc<EasyBlock>> {
+fn parse_encoded_blocks(bytes: &[u8]) -> HashSet<Arc<HardBlock>> {
     let mut b = HashSet::new();
     let rlp = Rlp::new(bytes);
 
     for slice in rlp.iter() {
-        b.insert(Arc::new(EasyBlock::from_bytes(slice.as_raw()).unwrap()));
+        b.insert(Arc::new(HardBlock::from_bytes(slice.as_raw()).unwrap()));
     }
 
     b
 }
 
-impl<'a> Chain<'a, EasyBlock, EasyBlockIterator<'a>> for EasyChain {
-    fn genesis(&self) -> Arc<EasyBlock> {
-        unimplemented!();
-    }
-
-    fn query(&self, hash: &Hash) -> Option<Arc<EasyBlock>> {
+impl<'a> Chain<'a, HardBlock, HardBlockIterator<'a>> for HardChain {
+    fn query(&self, hash: &Hash) -> Option<Arc<HardBlock>> {
         if let Some(cached) = self.block_cache.borrow_mut().get(hash) {
             Some(cached.clone())
         } else {
             if let Some(stored) = self.db.get(hash) {
                 // Store to heap and cache result
-                let heap_stored = Arc::new(EasyBlock::from_bytes(&stored).unwrap());
+                let heap_stored = Arc::new(HardBlock::from_bytes(&stored).unwrap());
                 self.block_cache
                     .borrow_mut()
                     .put(hash.clone(), heap_stored.clone());
@@ -194,7 +197,7 @@ impl<'a> Chain<'a, EasyBlock, EasyBlockIterator<'a>> for EasyChain {
         }
     }
 
-    fn query_by_height(&self, height: usize) -> Option<Arc<EasyBlock>> {
+    fn query_by_height(&self, height: usize) -> Option<Arc<HardBlock>> {
         unimplemented!();
     }
 
@@ -202,7 +205,7 @@ impl<'a> Chain<'a, EasyBlock, EasyBlockIterator<'a>> for EasyChain {
         unimplemented!();
     }
 
-    fn append_block(&mut self, block: Arc<EasyBlock>) -> Result<(), ChainErr> {
+    fn append_block(&mut self, block: Arc<HardBlock>) -> Result<(), ChainErr> {
         let top = &self.canonical_top;
 
         // The block must have a parent hash and the parent
@@ -249,18 +252,18 @@ impl<'a> Chain<'a, EasyBlock, EasyBlockIterator<'a>> for EasyChain {
         self.height
     }
 
-    fn canonical_top(&self) -> Arc<EasyBlock> {
+    fn canonical_top(&self) -> Arc<HardBlock> {
         self.canonical_top.clone()
     }
 
-    fn iter_canonical_tops(&'a self) -> EasyBlockIterator<'a> {
-        EasyBlockIterator(Box::new(
+    fn iter_canonical_tops(&'a self) -> HardBlockIterator<'a> {
+        HardBlockIterator(Box::new(
             self.canonical_tops_cache.iter().map(AsRef::as_ref),
         ))
     }
 
-    fn iter_pending_tops(&'a self) -> EasyBlockIterator<'a> {
-        EasyBlockIterator(Box::new(
+    fn iter_pending_tops(&'a self) -> HardBlockIterator<'a> {
+        HardBlockIterator(Box::new(
             self.pending_tops_cache.iter().map(AsRef::as_ref),
         ))
     }
